:recycle:  add type Maybe<T>
